Improved FileSystemInfoWrapper by using lazy loading
The file system type is only checked if the type is really needed
for setting the modification date correctly. And only if the modification
date is perhaps out of range of the possible underlaying file systems. Therefor
it is nearly never checked, instead of every time a new instance is created.